Added functions for jet background selection without any HLT

+++ SelectedX/header/FileMgr.h
@@ -4367,7 +4367,7 @@ void FileMgr::SetupChain(Int_t i_tuple, TChain *chain)
         cout << "No process set!" << endl;
         return;
     }
-    for (Int_t i_tup=0; i_tup<nNtuples.size(); i_tup++)
+    for (Int_t i_tup=0; i_tup<((int)(nNtuples.size())); i_tup++)
     {
         if (i_tup != i_tuple && i_tuple != -1) continue;
         for (Int_t i=1; i<=nNtuples[i_tup]; i++)
